Finish connecting to game and chat servers
Now it's sending the security token after connecting, and receiving back
a lot of messages that aren't handled so far.

Also replaced printf with iostream for printing out information, which
makes it easier to stream things like Message{In,Out} and to classify
things as information/warning or error. Not to mention std::endl
flushes.