More reorganizations, make it possible to choose character again
Of cource it doesn't really do anything yet, though. Still struggling
with how to give this thing shape in the context of QML.

A new class called Root was introduced which is used as the root context
object, so that all its properties are globally available. Its
properties include classes like the LoginManager (to be renamed to
AccountHandler) and GameHandler, which provide further interfaces and
properties. This is to avoid instantiating these classes from QML, which
looks like it could get messy.