Download and display the current map the player is on
Needs a small server patch to have it sent a base url for the data
(replaces the update url, but is sent in addition for backwards
compatibility).

The MapReader now has a lazy mode, where it will not load any external
tilesets or images, but still create the necessary Tile instances. The
images can then be loaded incrementally later.

Downloading happens threaded based on QNetworkAccessManager. Nothing is
cached at the moment.