Redid the Button and LineEdit components to be more suitable on mobiles
The screen resolution on a mobile is commonly much higher than that of a
desktop and is used with fat fingers. Though I might have overdone it a
bit, still needs to be tried on an actual device.

Also fixed the size of the logo in the background when using Qt 4.7.0,
by working around a small inconsistency in QML behaviour between that
version and current Qt.