Small protocol updates, fixing player identification
Fixed handling of GPMSG_BEING_ENTER, which had the being direction
added (ignored for now). This makes sure the character name is read
correctly again, which is used to identify the local player.

Emit a "kicked" signal when kicked by an administrator. Currently
unused.

Display an appropriate error message when trying to create a character
in an invalid slot (though character creation is currently not
possible).