Request properly resolved URLs
Now when the ResourceManager is asked to get

   maps/goldenfields/../../tiles/sand.png

It will use QUrl::resolve when combining it with the base URL, to get:

   http://data.sourceoftales.org/tiles/sand.png

In order to still be able to match replies to the requested files, a
user attribute is set on the QNetworkRequest, which was cleaner anyway.

This change was mostly needed because CloudFiles does not handle the
previously used non-clean URLs. It may be because a redirect is not
followed since it works in the browser, so support for redirects should
still be checked.