Wait for paths.xml to be loaded before continuing
To avoid characters being created too early, the different pages used
during the login process are now dynamically created. This should also
be more efficient since it avoids them eating any memory and possible
CPU during the game.

Dynamic creation of the pages does seem to perform better with Qt 5 than
with Qt 4. On Qt 4 the transition to the LoginPage is often lost in a
hiccup.