Implemented SceneGraph-based tile map rendering
Since the renderer can no longer rely on the 'exposed rectangle', the
QML side now informs the MapItem about the actually visible part of the
map. Each TileLayerItem child of the map uses this to only add
potentially visible tiles to the scene graph.

The tiles are rendered in batches as long as multiple tiles from the
same tileset follow each other. This should generally be quite
efficient. Further optimizations would be possible, especially for
layers without 'oversized' tiles where the rendering order does not
matter. Also, when a lot of tiles can be on the screen, it would make
sense to split up the map in chunks to avoid re-creating parts that
remain visible.