Removed ItemInfo::null
This should not be necessary. In QML the isNull property was nowhere
used, but anyway a comparison with null should work just fine.

On the C++ side, isNull was also not used, but in fact the returned
item is being checked against 0 for some debug warnings (which would
not have worked until now).

Finally, nobody owned this ItemInfo instance, and passing it to QML
would have risked it being deleted by the garbage collector.