Made the player name and chat messages appear on top of the map
They shouldn't be hiding behind the trees. :-)

To do this the parent is set to be the map, and the stacking order is
set to 65537, which places it on top of the layers that are above
the sprites (which live at 65536). I realize a more explicit layering
may be in order, but this works for now.