Added a New Character page
It is now possible to create new characters. Only the name and hair
style are configurable for now.

The 'Character' can now be created in QML and the gender, hair style and
hair color can be modified from QML. This allows re-using the same
display method used for in-game characters on the New Character page.