Use more accurate timing in SpriteItem
Using QTimer was not very accurate and often resulted in sprite frames
not matching up. With the new method there's still a small chance of
this happening, but it's very rare now.

A SpriteItem that is playing an animation is now continually causing its
updatePaintNode function to be called, which advances the animation
based on a QElapsedTimer.

Also, the SpriteItem will now be at complete rest when the animation has
only a single frame (generally the case for standing animations).