Send messages immediately by default
enet_peer_send actually only enqueues a packet, the sending happens on
the next call to enet_host_service. Since we service the hosts only
every 100ms, we were adding up to 100ms of delay to message delivery.

The client only rarely sends messages to the server, so it makes sense
to send them immediately by default. This is now done by calling
enet_host_flush after sending. For cases where this is not desired, a
new method ENetClient::enqueue was introduced. This new method is
currently only used for the frequent PGMSG_WALK messages.