Fixed handling of animations with a single endless frame
Since an endless frame has duration -1, if it is the only frame in an
animation the whole animation will have duration -1. Such animations
were not playing at all.

At the same time, frames with duration 0 are dropped while reading
animations to avoid the SpriteItem from looping endlessly in search of a
frame to display.