Make Being::lookAt prefer up/down a little more
Especially on Android, monsters and NPCs could be seen changing
direction very frequently when walking diagonally. This probably has to
do with lower precision due to qreal being a 'float' rather than a
'double' on ARM.

Fixed it by preferring the up/down directions when there is only a
slight (less than 10%) difference between the vertical and horizontal
walking speeds.

Of course, this would create the same frequent direction changes when
an entity moves at exactly the new angle. Fortunately that doesn't
happen.