Changed resource manager to prefer loading from cache
Mainly to test how that performs and to check whether we can live with
regular HTTP expiration settings.

Eventually it would be nice to switch to "never expire" in combination
with manually removing changed files from the cache, but I hope we can
do without such a system for now.