Fixed drawing order of tiles on the fringe layer
Now they correctly interleave with the entities rather than always being
drawn below them.

To achieve this, they each get their own QQuickItem (a TileItem) so that
the right z value can be set on them. The items are only created for
actually visible fringe tiles. For a large number of fringe tiles this
could be further optimized by having only an item for each row.