Removed a bunch of unnecessary unload() calls
The AbilityDB, HairDB, NpcDB and RaceDB do not need to do a manual
unloading upon destruction, because the objects that they manage are
their children and as such will already be automatically cleaned up.

MonsterDB was failing to clean up its MonsterInfo instances entirely,
though when some would be returned from getInfo to the QmlEngine, it
would have taken ownership of them since they didn't have any owner (but
this currently never happens).

Actually I was trying to fix a crash on exit with Qt 5.1.1, but this is
still there. Fortunately, it seems to be more a problem in Qt which has
been fixed in Qt 5.2 beta1.